refact

+++ bcoins/coins.go
@@ -13,7 +13,7 @@ type Coin struct {
 	Insight       *Insight
 	TxVersion     int32
 	MinRelayTxFee float64
-	FeePerByte int64
+	FeePerByte int
 }
 
 type Insight struct {
@@ -31,11 +31,11 @@ type Network struct {
 
 var coins = map[string]Coin{
 	"via": {Name: "viacoin", Symbol: "via", Network: &Network{Name: "viacoin", P2PKH: 0x47, P2SH: 0x21, PrivateKeyID: 0xC7, magic: 0xcbc6680f},
-		Insight: &Insight{"https://explorer.viacoin.org", "https://explorer.viacoin.org/api"}, TxVersion: 2, MinRelayTxFee: 0.001, FeePerByte:int64(110),
+		Insight: &Insight{"https://explorer.viacoin.org", "https://explorer.viacoin.org/api"}, TxVersion: 2, MinRelayTxFee: 0.001, FeePerByte:110,
 	},
 
 	"ltc": {Name: "litecoin", Symbol: "ltc", Network: &Network{Name: "litecoin", P2PKH: 0x30, P2SH: 0x05, PrivateKeyID: 0xb0, magic: 0xfbc0b6db},
-		Insight: &Insight{"https://explorer.viacoin.org", "https://explorer.viacoin.org/api"}, TxVersion: 2, MinRelayTxFee: 0.001, FeePerByte:int64(280),
+		Insight: &Insight{"https://explorer.viacoin.org", "https://explorer.viacoin.org/api"}, TxVersion: 2, MinRelayTxFee: 0.001, FeePerByte:280,
 	},
 }
